Added item pick up system.
Added dropped item entity that currently stores the path to an inventory item. Temporary texture is the gem.
Skeleton wizard drops a firestaff that randomizes which type of projectile it fires. (laser is omitted for now, needs changed)
player can pick up items with the "Y map" (Z on keyboard).